HYD-4606 Add a "some" method to the fp module
A useful function in the fp library is the ability to define a function,
run that function against a set of data and see if the result of running
that function returns true at least once. This abstraction has many uses
and thus it will prove to be beneficial to our fp library.

This patch will update the fp module by adding the "some" function as
well as a set of unit tests.

  - Updated fp index file to include the some function
  - Updated the fp module tests with a new section to test the some
  function

Signed-off-by: Will Johnson <william.c.johnson@intel.com>