HYD-4748 FP: add head method
While working on a parser, it's quite useful to pull the first value off an array of tokens.
Add fp.head method which will return the first value of an array.
This will work with strings as well.

  - Add fp.head.
  - Add tests.

HYD-4749 FP: Add toArray method

While working on a parser, it's quite useful to return a consumed substream of tokens.
If we have consumed a single token, we still need to wrap it back to an array.
Add fp.toArray method which will wrap a passed value in an array.

  - Add fp.toArray.
  - Add tests.

HYD-4750 FP: Add either method

When working with a parser, there could be an error parsing an array of tokens.
It is useful to have a way to stop on error and abort parsing the rest of the tokens.
Create a simple representation of Either
(example: https://hackage.haskell.org/package/category-extras-0.53.4/docs/Control-Monad-Either.html).
Instead of creating Left and Right types, we will use Error as a left. Everything else is a right.

  - Add fp.either.
  - Add tests.

Signed-off-by: Joe Grund <joe.grund@intel.com>