HYD-4827 fp@1.0.7: Add reduce method
Add reduce to fp. reduce works much the same as Array.prototype.reduce,
it takes an initial value, an object to reduce, and a reduce function
and returns a transformed value.

The typical example is reducing a list of numbers to a single sum.

This method will differ in that it will delegate to the reduce method on
any obj passed in if available. It also will reduce a singular value.

  - Add reduce method.
  - Add tests.

Signed-off-by: Joe Grund <joe.grund@intel.com>