HYD-4833 Create mapFn function to span a finite set of arguments across n functions
We've seen a pattern in which a finite set of arguments is being called
across n functions. Take this simple case:
translate(xScale(name), yScale(name));

In this case, we are passing the same argument (name) to xScale and
yScale. By adding a mapFp function this can be simplified to:

var translate = fp.flow(fp.mapFn(xScale, yScale),
fp.invoke('translate'));
translate(name);

This allows us to use a more functional approach. mapFn will essentially
invoke each function with the arguments being passed in. It will return
an array of the results.

Signed-off-by: Will Johnson <william.c.johnson@intel.com>