HYD-4834 Create tap method in fp library
Sometimes we need to perform an operation that is a side effect. Take
for example:

fp.flow(canFit, fp.eq(false), function (coords)
{
    labels.attr('display', 'none');
    return mapToCoords(circleLens,coords);
})

In this case, if the context is determined that it cannot fit, we
update the display attribute to none such that the labels are hidden
and then map the coordinates using the width of the circle. The
reason we are having to do these two operations inside an anonymous
function is because updating the labels is a side effect. In other
words, there is no input or output, it just needs to happen before
we can re-map the coordinates. This is where using the tap function
becomes helpful. It would be nicer to do the following:

fp.flow(fp.tap(labels.attr.bind(null, ‘display’, ‘none’),
mapToCoords(circleLens));

In this case, the inputs are simply passed through and reach
mapToCoords.

Signed-off-by: Will Johnson <william.c.johnson@intel.com>