HYD-4910 Create wrapArgs method
I ran across a case while using fp in which I have multiple arguments
being passed to a flow. The problem is that flow only takes one arg. The
only way to work points free in this case is to wrap the list of args
into an array and pass it into flow.

This patch will update fp such that an wrapArgs method will be added.

Signed-off-by: Will Johnson <william.c.johnson@intel.com>