fp: Add invokeMethodN HYD-5279 fp: create invokeMethodN
We currently can invoke methods in a functional way using
fp.invokeMethod. This works well, but there are cases where we'd like to
invoke the result as part of another pipeline. Currently,
fp.invokeMethod takes an array to represent the args which causes us to
have to perform extra wrapping of args before calling the function.

Instead we could curry the resulting function to a given arity
beforehand. If we do this, we can use the resulting function without
having to do any wrapping beforehand. Implement fp.invokeMethodN to do
this.

  - Implement fp.invokeMethodN + tests.
  - Bump deps.
  - Add some more type annotations.

Signed-off-by: Joe Grund <joe.grund@intel.com>