HYD-5349 fp: create eqProp
There is a pattern we repeat quite often, which is to compare the
property of an object for equality to some value. It looks like this:

eqFn(lensProp(prop), identity, a, b)

We should raise this to it's own specialized method eqProp. It will look
like:

eqProp('prop', a, b);

  - implement eqProp + tests.
  - updated deps.

Signed-off-by: Joe Grund <joe.grund@intel.com>