HYD-5632 FP: Allow cond to return "empty" values
I ran across an issue while using cond in which a satisfied condition
will not return an "empty" value, such as an empty string. The cond
method should receive an array of corresponding predicate and evaluation
function pairs. The evaluation function should return the result if the
predicate passes.

This patch will update cond such that if the predicate returns true the
result of the evaluation function will return, even if it evaluates to
an "empty" value, such as an empty string.

- Updated the inner cond to set the value on result and then return true
  as opposed to returning the actual result. This ensures that if the
  predicate results in true the result will be the result of passing the
  value (xs) into the evaluation function, even if that result is
  considered an empty value.
- Updated tests to verify that empty values can be returned.

Signed-off-by: Will Johnson <william.c.johnson@intel.com>