HYD-5723 fp: add memoize method
Add a memoize method to fp. This will mostly be useful in allowing us to
use Angular 1 filters in a similar manner to how Angular 2 uses pipes.

In short, it will allow us to bail out of code execution if our input
arguments match a value that has already been executed before.

  - Add memoize method + tests.
  - Bump deps to latest.
  - Update dotfiles.
  - Add shrinkwrap.
  - Update code to conform to latest eslint / jscs rules.

Signed-off-by: Joe Grund <joe.grund@intel.com>