HYD-5901 fp: add uniqBy method
We need a way to determine uniqueness of a set of values. The built-in
Set is too primitive as it only compares using identity, and does not
provide a way to focus on a subset of a value for uniqueness.

This method will use indexOf internally to determine if a value is a
member of a set.

  - Implement uniqBy method + tests.
  - Bump deps + shrinkwrap.
  - Add plugin to enforce flow annotation styles.

Signed-off-by: Joe Grund <joe.grund@intel.com>