HYD-6749 Simplify FP
The `fp` library was originally intended to be polymorphic
at the method level.

I.E. if there was an object that had a delegate, we would call the
delegate.

The current downside is that we cannot statically check polymorphic
curried functions if
we don't know the container type ahead of time (which is a contradiction
in itself).

As an example, consider a curried `map` which could take any object
containing a map method.

```
fp.map(x => x + 1, [1, 2, 3]);
fp.map(x => x + 1, highland([1, 2, 3]);
```

Both work fine because we can tell what the container is.

Now consider

```
const adder = fp.map(x => x + 1);
```

This is a polymorphic map, it delegates to any
container and adds 1 to it's value(s).

The problem is flow cannot infer what that container is because
we have not provided it yet, (any array does not conform to the functor
interface).

The solution is to consider functions polymorphic (`x => x + 1`)
and container methods tied to a container.

This means fp will become specialized to arrays,
and other data type modules will contain their own
specialization of map.

Finally, we will remove placeholder support. It was rarely used in
practice and adds considerable complexity to writing types.

  - Remove type level delegate support, which was broken.
  - Remove placeholder from curry.
  - Update tests.
  - Update deps.
  - Add last method to cover what tail used to do.
  - Replace void with empty.

Signed-off-by: Joe Grund <joe.grund@intel.com>