HYD-7128 FP: Add match function
Adding a match function will make for easier chaining (and other operations) in task-either and other libraries. It will allow us to map based on a specific type, resulting in cleaner code. Take this example with Task-Either:

```
fileExists(path)
      .chainLeft(
        match([[FileDoesNotExist, x => ofRight(x)], [NotAFile, x => ofLeft(x)]])
      )
      .chain(
        match([
          [FileExists, () => getFileHash(path)],
          [FileDoesNotExist, () => ofRight('')]
        ])
      );
```

In this code, it is easy to see what is happening. When chaining, we get the fileHash if the file exists, an empty string if the file doesn't exist, and pass along the entry if it exists but is not a file. This is a primitive version of match as it is not as expressive as other ML languages, but it will allow us to get the job done.

Signed-off-by: Will Johnson <william.c.johnson@intel.com>