Update to latest (#3)
* HYD-6749 Simplify FP

The `fp` library was originally intended to be polymorphic
at the method level.

I.E. if there was an object that had a delegate, we would call the
delegate.

The current downside is that we cannot statically check polymorphic
curried functions if
we don't know the container type ahead of time (which is a contradiction
in itself).

As an example, consider a curried `map` which could take any object
containing a map method.

```
fp.map(x => x + 1, [1, 2, 3]);
fp.map(x => x + 1, highland([1, 2, 3]);
```

Both work fine because we can tell what the container is.

Now consider

```
const adder = fp.map(x => x + 1);
```

This is a polymorphic map, it delegates to any
container and adds 1 to it's value(s).

The problem is flow cannot infer what that container is because
we have not provided it yet, (any array does not conform to the functor
interface).

The solution is to consider functions polymorphic (`x => x + 1`)
and container methods tied to a container.

This means fp will become specialized to arrays,
and other data type modules will contain their own
specialization of map.

Finally, we will remove placeholder support. It was rarely used in
practice and adds considerable complexity to writing types.

  - Remove type level delegate support, which was broken.
  - Remove placeholder from curry.
  - Update tests.
  - Update deps.
  - Add last method to cover what tail used to do.
  - Replace void with empty.

Signed-off-by: Joe Grund <joe.grund@intel.com>

* Version 6.0.0

eccf336 HYD-6749 Simplify FP

* HYD-6752 Return Maybe in place of a void union.

We currently type things like fp.find as

```
declare export function find<A>(fn:Fn1<A, boolean>, xs:Array<A>,
...rest:empty[]): A | void;
```

This means the return type could be void. Instead; model the void choice
as a maybe. The signature will then look like:

```
declare export function find<A>(fn:Fn1<A, boolean>, xs:Array<A>,
...rest:empty[]): Maybe<A>;
```

  - Update find, head, last to return Maybe.
  - Update curry2 to work with circular imports.

Signed-off-by: Joe Grund <joe.grund@intel.com>

* Version 7.0.0

97d1806 HYD-6752 Return Maybe in place of a void union.

* HYD-7007 Update fp to use iml scope

Our Artifactory recently updated to support scoped packages.
Update fp to use the iml scope.

  - Update scope.
  - Switch to jest.
  - Only support node 6.
  - Remove dead code for tap.
  - add yarn.

Signed-off-by: Joe Grund <joe.grund@intel.com>

* Version 7.0.1

191ef01 HYD-7007 Update fp to use iml scope

* remove more stuff from build

Signed-off-by: Joe Grund <joe.grund@intel.com>

* Version 7.0.2

a8f424a remove more stuff from build

* Version 7.0.3

6aaa7a7 update lockfile

* ignore dist/test

Signed-off-by: Joe Grund <joe.grund@intel.com>

* Version 7.0.4

3e405b0 ignore dist/test

* HYD-7033 FP: bump deps

Bump dependencies in fp to latest.

Dependencies bumped are as follows.

"babel-preset-env": "1.1.7", -> "babel-preset-env": "1.1.8"
"eslint": "3.13.1" -> "eslint": "3.13.1"
"eslint-plugin-flowtype": "2.29.2" -> "eslint-plugin-flowtype": "2.30.0"
"eslint-plugin-import": "2.2.0" ->		     "eslint-plugin-import": "2.2.0"
"flow-bin": "0.37.4" -> "flow-bin": "0.38.0"

Signed-off-by: Joe Grund <joe.grund@intel.com>

* HYD-7046 fp: stop compiling to ES5

Stop compiling maybe to ES5. We can use downstream repos to do this.

Signed-off-by: Joe Grund <joe.grund@intel.com>

* 7.0.5

* fixup source

Signed-off-by: Joe Grund <joe.grund@intel.com>

* Version 7.0.6

2586d12 fixup source

* update a few missed import changes

Signed-off-by: Joe Grund <joe.grund@intel.com>

* bump to 7.0.7

Signed-off-by: Joe Grund <joe.grund@intel.com>

* Update typings #39

flow 0.39 changed the way overloaded types are used.

Because of this, we need to switch ordering from least specific to most specific to get tests passing.
While we're here, update deps and implement prettier.

  - Update types.
  - Implement prettier.
  - Bump deps.

Signed-off-by: Joe Grund <joe.grund@intel.com>

* 7.0.8

* HYD-7128 FP: Add match function

Adding a match function will make for easier chaining (and other operations) in task-either and other libraries. It will allow us to map based on a specific type, resulting in cleaner code. Take this example with Task-Either:

```
fileExists(path)
      .chainLeft(
        match([[FileDoesNotExist, x => ofRight(x)], [NotAFile, x => ofLeft(x)]])
      )
      .chain(
        match([
          [FileExists, () => getFileHash(path)],
          [FileDoesNotExist, () => ofRight('')]
        ])
      );
```

In this code, it is easy to see what is happening. When chaining, we get the fileHash if the file exists, an empty string if the file doesn't exist, and pass along the entry if it exists but is not a file. This is a primitive version of match as it is not as expressive as other ML languages, but it will allow us to get the job done.

Signed-off-by: Will Johnson <william.c.johnson@intel.com>

* - Removed types from source

* - Try adding types to match tests. Don't have it working yet.

* - Added a declaration for each match permutation
- Added type tests
- Added tests for each permutation

* - Updated declarations to account for 5 permutations
- Added permutations for type tests
- Added permutations for unit tests

* - More type updates

* - Updating types and tests. Still have some broken tests that need to be
  fixed.

* - Fix test errors
- Updated implementation to account for the three types of input
  variations
- Added an "isClass" function

* - Update isClass use toString on the object being passed in instead of
  Function.prototype
- Move test functions outside of expect block
- Add types to callbacks in test

* fixup styles

Signed-off-by: Joe Grund <joe.grund@intel.com>

* 7.1.0

* bump-maybe

Signed-off-by: Joe Grund <joe.grund@intel.com>

* HYD-7141 fp: Remove use of curry and replace with binding

Curry is currently giving us false positives for our type system. Curry
needs to be removed and replaced with binding. This ticket will replace
all fp functions that use curry with binding.

Signed-off-by: Will Johnson <william.c.johnson@intel.com>

* 8.0.0

* update to latest

Signed-off-by: Joe Grund <grundjoseph@gmail.com>